refactor: Standardize monetary and quantity fields to integer types and boolean flags in models, controllers, and API documentation.

+++ database/migrations/2026_01_13_150838_create_transactions_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,9 +14,9 @@ return new class extends Migration
             $table->id();
             $table->string('transaction_code')->unique();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->decimal('total_amount', 12, 2);
-            $table->decimal('discount_amount', 12, 2)->default(0);
-            $table->decimal('final_amount', 12, 2);
+            $table->integer('total_amount');
+            $table->integer('discount_amount')->default(0);
+            $table->integer('final_amount');
             $table->string('payment_method');
             $table->string('status')->default('completed');
             $table->text('notes')->nullable();
